split four classes in namespace formatter to different files
