+++ source/module_base/CMakeLists.txt
@@ -48,7 +48,10 @@ add_library(
     parallel_reduce.cpp
     spherical_bessel_transformer.cpp
     cubic_spline.cpp
-    formatter.cpp
+    formatter_fmt.cpp
+    formatter_physfmt.cpp
+    formatter_table.cpp
+    formatter_contextfmt.cpp
     ${LIBM_SRC}
 )
