+++ source/module_base/test/CMakeLists.txt
@@ -183,6 +183,6 @@ AddTest(
 
 AddTest(
   TARGET formatter_test
-  SOURCES ../formatter.cpp formatter_fmt_test.cpp formatter_physfmt_test.cpp formatter_table_test.cpp formatter_contextfmt_test.cpp
-  #LIBS formatter
+  SOURCES ../formatter_fmt.cpp ../formatter_physfmt.cpp ../formatter_table.cpp ../formatter_contextfmt.cpp 
+          formatter_fmt_test.cpp formatter_physfmt_test.cpp formatter_table_test.cpp formatter_contextfmt_test.cpp
 )
