+++ source/module_base/test/formatter_contextfmt_test.cpp
@@ -1,4 +1,4 @@
-#include "../formatter.h"
+#include "../formatter_contextfmt.h"
 #include <gtest/gtest.h>
 
 /************************************************
