+++ source/module_base/test/formatter_physfmt_test.cpp
@@ -1,4 +1,4 @@
-#include "../formatter.h"
+#include "../formatter_physfmt.h"
 #include <gtest/gtest.h>
 
 /************************************************
