+++ source/module_md/test/CMakeLists.txt
@@ -31,7 +31,10 @@ list(APPEND depend_files
   ../../module_base/global_variable.cpp
   ../../module_base/libm/branred.cpp
   ../../module_base/libm/sincos.cpp
-  ../../module_base/formatter.cpp
+  ../../module_base/formatter_fmt.cpp
+  ../../module_base/formatter_physfmt.cpp
+  ../../module_base/formatter_table.cpp
+  ../../module_base/formatter_contextfmt.cpp
   ../../module_cell/module_neighbor/sltk_adjacent_set.cpp
   ../../module_cell/module_neighbor/sltk_atom_arrange.cpp
   ../../module_cell/module_neighbor/sltk_atom.cpp
